requirements.txt
Flask
twilio
python-dotenv
gunicorn
requests
openai
pymongo
pinecone
redis